--- lib/ad_localize/mappers/ios_translation_mapper.rb
@@ -1,12 +1,0 @@
-module AdLocalize
-  module Mappers
-    class IOSTranslationMapper < TranslationMapper
-      private
-
-      def sanitize_value(value:)
-        return if value.blank?
-        value.gsub(/(?<!\\)\"/, "\\\"")
-      end
-    end
-  end
-end
