+++ lib/ad_localize/serializers/info_plist_serializer.rb
@@ -12,16 +12,8 @@ module AdLocalize
       end
 
       def variable_binding(locale_wording:)
-        translations = locale_wording.info_plists.map { |translation| translation_to_binding(translation:) }
+        translations = locale_wording.info_plists.map { |translation| map_simple_wording(translation:) }
         { translations: translations }
-      end
-
-      def translation_to_binding(translation:)
-        SimpleWordingViewModel.new(
-          label: translation.key.label,
-          value: sanitize_value(value: translation.value),
-          comment: translation.comment
-        )
       end
 
       def sanitize_value(value:)
