+++ lib/ad_localize/serializers/localizable_stringsdict_serializer.rb
@@ -18,20 +18,6 @@ module AdLocalize
         }
       end
 
-      def map_compound_wording(label:, translations:)
-        variants = translations.map { |translation| translation_to_binding(translation:) }
-        CompoundWordingViewModel.new(label:, variants:)
-      end
-
-      def translation_to_binding(translation:)
-        SimpleWordingViewModel.new(
-          label: translation.key.label,
-          value: sanitize_value(value: translation.value),
-          comment: translation.comment,
-          variant_name: translation.key.variant_name
-        )
-      end
-
       def sanitize_value(value:)
         return if value.nil? || value.strip.empty?
         value.gsub(/(?<!\\)\"/, "\\\"")
