+++ lib/ad_localize/serializers/properties_serializer.rb
@@ -5,22 +5,13 @@ module AdLocalize
 
       private
 
-      def translation_to_binding(translation:)
-        SimpleWordingViewModel.new(
-          label: translation.key.label,
-          value: sanitize_value(value: translation.value),
-          comment: translation.comment,
-          variant_name: translation.key.variant_name
-        )
-      end
-
       def template_path
         TEMPLATES_DIRECTORY + "/properties/template.properties.erb"
       end
 
       def variable_binding(locale_wording:)
         {
-          translations: locale_wording.singulars.map { |translation| translation_to_binding(translation:) }
+          translations: locale_wording.singulars.map { |translation| map_simple_wording(translation:) }
         }
       end
