+++ lib/ad_localize/serializers/with_template.rb
@@ -1,14 +1,35 @@
 module AdLocalize
   module Serializers
     module WithTemplate
+      # TODO: add sanitizers in DI them in serializer so that sanitize value can be common
+      private
+      
       TEMPLATES_DIRECTORY = __dir__ + "/../templates"
+
+      def template_path
+        raise 'Override me!'
+      end
 
       def variable_binding(locale_wording:)
         raise 'override me!'
       end
 
-      def translation_to_binding(translation:)
-        raise 'override me!'
+      def map_simple_wording(translation:)
+        SimpleWordingViewModel.new(
+          label: translation.key.label,
+          value: sanitize_value(value: translation.value),
+          comment: translation.comment,
+          variant_name: translation.key.variant_name
+        )
+      end
+
+      def sanitize_value(value:)
+        raise 'Override me!'
+      end
+
+      def map_compound_wording(label:, translations:)
+        variants = translations.map { |translation| translation_to_binding(translation:) }
+        CompoundWordingViewModel.new(label:, variants:)
       end
 
       def render(locale_wording:)
